switch fresh async render trigger to an island
