+++ hooks/mod.ts
@@ -1,7 +1,4 @@
 export { useDevice } from "./useDevice.ts";
-export {
-  useAsyncRenderAttributes,
-  usePartialSection,
-} from "./usePartialSection.ts";
+export { usePartialSection } from "./usePartialSection.ts";
 export { useScript, useScriptAsDataURI } from "./useScript.ts";
 export { useSection } from "./useSection.ts";
