+++ runtime/fresh/Bindings.tsx
@@ -3,46 +3,8 @@
 
 import { Head, Partial } from "$fresh/runtime.ts";
 import type { Framework } from "@deco/deco";
-import { useAsyncRenderAttributes, usePartialSection } from "@deco/deco/hooks";
-
-const script = (id: string) => {
-  function init() {
-    const elem = document.getElementById(id);
-    const parent = elem?.parentElement;
-
-    if (elem == null || parent == null) {
-      console.error(
-        `Missing element of id ${id} or its parent element. Async rendering will NOT work properly`,
-      );
-      return;
-    }
-
-    const observeAndClose = (e: IntersectionObserverEntry[]) => {
-      e.forEach((entry) => {
-        if (entry.isIntersecting) {
-          elem.click();
-          observer.disconnect();
-        }
-      });
-    };
-    const observer = new IntersectionObserver(observeAndClose);
-    observer.observe(parent);
-    observeAndClose(observer.takeRecords());
-  }
-
-  if (document.readyState === "complete") {
-    init();
-  } else {
-    addEventListener("load", init);
-  }
-};
-
-const dataURI = (fn: typeof script, id: string) =>
-  btoa(
-    `decodeURIComponent(escape(${
-      unescape(encodeURIComponent(`((${fn})("${id}"))`))
-    }))`,
-  );
+import { usePartialSection } from "@deco/deco/hooks";
+import DispatchAsyncRender from "./islands/DispatchAsyncRender.tsx";
 
 const bindings: Framework = {
   name: "fresh",
@@ -60,16 +22,12 @@ const bindings: Framework = {
           <link rel="prefetch" href={href} as="document" />
         </Head>
         <button
-          {...useAsyncRenderAttributes()}
           f-partial={href}
           {...rest}
           id={btnId}
           style={{ display: "none" }}
         />
-        <script
-          defer
-          src={`data:text/javascript;base64,${dataURI(script, btnId)}`}
-        />
+        <DispatchAsyncRender id={btnId} />
         {children}
       </>
     );
